pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "vision-workspace"
version = "0.1.0"
description = "Utilities and workflows for the vision workspace."
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
  "ultralytics>=8.0.0",
  "opencv-python>=4.8.0",
  "numpy>=1.23",
  "pillow>=10.0.0",
  "pyyaml>=6.0.0",
  "tqdm>=4.66.0",
  "onnx>=1.15.0",
  "onnxruntime>=1.16.0",
]

[project.scripts]
init-project = "vision_workspace.init_project:main"
create-dataset = "vision_workspace.create_dataset:main"
check-environment = "vision_workspace.check_environment:main"
visualize-labels = "vision_workspace.visualize_labels:main"
init-experiment = "vision_workspace.init_experiment:main"
train-model = "vision_workspace.train_model:main"
evaluate-experiment = "vision_workspace.evaluate_experiment:main"
run-inference = "vision_workspace.run_inference:main"
export-onnx = "vision_workspace.export_onnx:main"



[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
